requirements.txt
# Uses Python 3.10.10

# Base libraries
matplotlib==3.9.2
numpy==1.23.2
scikit-learn==1.2.0
pandas
scipy
seaborn
tqdm==4.64.0
cvxpy==1.5.2

pydash==7.0.6
python-dotenv==1.0.0

kaggle==1.5.12

# Pretty logs
rich==13.4.1

# Code style and type-checking
black==22.3.0
flake8==4.0.1
isort==5.12.0
pytest==7.1.2
pylint==2.13
mypy==1.2
pre-commit==3.2.2 # Run `pre-commit install` to set-up pre-commit hooks

# Cooper -- Constrained optimization
git+https://github.com/cooper-org/cooper.git@9f94213a74a903220f94dd0af6955129e5fdb4a0

# Pytorch
--extra-index-url https://download.pytorch.org/whl/cu124
torch==2.4.0

--extra-index-url https://download.pytorch.org/whl/cu124
torchvision==0.19.0

--extra-index-url https://download.pytorch.org/whl/cu124
torchaudio==2.4.0

# Experiment tracking and reproducibility
ml-collections==0.1.1
wandb==0.16.3

# Dimensionality reduction
umap-learn==0.5.3

# Submitting SLURM jobs
submitit==1.5.1

# Adversarial robustness
foolbox==3.3.3

# Dataset
wilds
